Okay, I'm pushing this so you can see arch.
I haven't tested anything, still working on it.
It shouldn't affect anything you were doing though.

A client that is viewing needs to find the
server's livestream object via the new map passed
in to each thread handler. Once he has
the livestream object, the client needs
to register for updates to that livestream.

Then, on each update the clientviewing thread
on the server simply sends the next segment
over to the actual client doing the viewing.

I just said the heck with it, we can use
serializable classes to make sending files
over the inter-webs easier :)