Got a full stream going for a single client single streamer.
You need to use the stream name "ina" for viewing ad starting a stream. We need to now work on cleaning the application up for handling closed connections, removing test code etc..

If you want to try it out, start the server, then a login client, select a file after clicking stream using the name "ina". Then run the watch client and start the stream with "ina" in the text field. You have to the maximize the window also.