Fixing some stuff.
Biggest problem right now is dealing with duplicate users. We might just
chuck error correction.

If I want to start a stream and I enter a name already present, it
becomes tedious network wise to keep checking until we get it correct.

If you can find a nice way to do it, feel free to.